+++ rascal-lsp/src/main/java/org/rascalmpl/vscode/lsp/BaseLanguageServer.java
@@ -256,7 +256,7 @@ public abstract class BaseLanguageServer {
                 logger.info("LSP connection started (connected to {} version {})", params.getClientInfo().getName(), params.getClientInfo().getVersion());
                 logger.debug("LSP client capabilities: {}", params.getCapabilities());
                 final InitializeResult initializeResult = new InitializeResult(new ServerCapabilities());
-                lspDocumentService.initializeServerCapabilities(initializeResult.getCapabilities());
+                lspDocumentService.initializeServerCapabilities(params.getCapabilities(), initializeResult.getCapabilities());
                 lspWorkspaceService.initialize(params.getCapabilities(), params.getWorkspaceFolders(), initializeResult.getCapabilities());
                 logger.debug("Initialized LSP connection with capabilities: {}", initializeResult);
                 return initializeResult;