+++ rascal-lsp/src/main/java/org/rascalmpl/vscode/lsp/parametric/LanguageContributionsMultiplexer.java
@@ -38,6 +38,7 @@ import org.rascalmpl.vscode.lsp.util.concurrent.CompletableFutureUtils;
 import org.rascalmpl.vscode.lsp.util.concurrent.InterruptibleFuture;
 
 import io.usethesource.vallang.IConstructor;
+import io.usethesource.vallang.IInteger;
 import io.usethesource.vallang.IList;
 import io.usethesource.vallang.ISet;
 import io.usethesource.vallang.ISourceLocation;
@@ -71,6 +72,7 @@ public class LanguageContributionsMultiplexer implements ILanguageContributions 
     private volatile CompletableFuture<ILanguageContributions> didRenameFiles = failedInitialization();
     private volatile CompletableFuture<ILanguageContributions> selectionRange = failedInitialization();
     private volatile CompletableFuture<ILanguageContributions> callHierarchy = failedInitialization();
+    private volatile CompletableFuture<ILanguageContributions> completion = failedInitialization();
 
     private volatile CompletableFuture<Boolean> hasAnalysis = failedInitialization();
     private volatile CompletableFuture<Boolean> hasBuild = failedInitialization();
@@ -87,6 +89,7 @@ public class LanguageContributionsMultiplexer implements ILanguageContributions 
     private volatile CompletableFuture<Boolean> hasDidRenameFiles = failedInitialization();
     private volatile CompletableFuture<Boolean> hasSelectionRange = failedInitialization();
     private volatile CompletableFuture<Boolean> hasCallHierarchy = failedInitialization();
+    private volatile CompletableFuture<Boolean> hasCompletion = failedInitialization();
 
     private volatile CompletableFuture<Boolean> specialCaseHighlighting = failedInitialization();
 
@@ -164,6 +167,7 @@ public class LanguageContributionsMultiplexer implements ILanguageContributions 
         didRenameFiles = findFirstOrDefault(ILanguageContributions::hasDidRenameFiles);
         selectionRange = findFirstOrDefault(ILanguageContributions::hasSelectionRange);
         callHierarchy = findFirstOrDefault(ILanguageContributions::hasCallHierarchy);
+        completion = findFirstOrDefault(ILanguageContributions::hasCompletion);
 
         hasAnalysis = anyTrue(ILanguageContributions::hasAnalysis);
         hasBuild = anyTrue(ILanguageContributions::hasBuild);
@@ -180,6 +184,7 @@ public class LanguageContributionsMultiplexer implements ILanguageContributions 
         hasDidRenameFiles = anyTrue(ILanguageContributions::hasDidRenameFiles);
         hasSelectionRange = anyTrue(ILanguageContributions::hasSelectionRange);
         hasCallHierarchy = anyTrue(ILanguageContributions::hasCallHierarchy);
+        hasCompletion = anyTrue(ILanguageContributions::hasCompletion);
 
         // Always use the special-case highlighting status of *the first*
         // contribution (possibly using the default value in the Rascal ADT if
@@ -336,13 +341,18 @@ public class LanguageContributionsMultiplexer implements ILanguageContributions 
     }
 
     @Override
-    public CompletableFuture<Boolean> hasSelectionRange() {
-        return hasSelectionRange;
+    public InterruptibleFuture<IList> selectionRange(IList focus) {
+        return flatten(selectionRange, c -> c.selectionRange(focus));
     }
 
     @Override
-    public InterruptibleFuture<IList> selectionRange(IList focus) {
-        return flatten(selectionRange, c -> c.selectionRange(focus));
+    public InterruptibleFuture<IList> completion(IList focus, IInteger cursorOffset, IConstructor trigger) {
+        return flatten(completion, c -> c.completion(focus, cursorOffset, trigger));
+    }
+
+    @Override
+    public CompletableFuture<IList> completionTriggerCharacters() {
+        return completion.thenCompose(ILanguageContributions::completionTriggerCharacters);
     }
 
     @Override
@@ -411,6 +421,11 @@ public class LanguageContributionsMultiplexer implements ILanguageContributions 
     }
 
     @Override
+    public CompletableFuture<Boolean> hasSelectionRange() {
+        return hasSelectionRange;
+    }
+
+    @Override
     public CompletableFuture<Boolean> hasRename() {
         return hasRename;
     }
@@ -420,9 +435,13 @@ public class LanguageContributionsMultiplexer implements ILanguageContributions 
         return hasDidRenameFiles;
     }
 
-    @Override
     public CompletableFuture<Boolean> hasCallHierarchy() {
         return hasCallHierarchy;
+    }
+
+    @Override
+    public CompletableFuture<Boolean> hasCompletion() {
+        return hasCompletion;
     }
 
     @Override